Default bare int literals to builtins.int, not i32; fix gaps this exposed
A bare, unannotated int literal (`x = 0`) previously defaulted to i32.
Per user direction, it now sugars into a real int(literal) construction
(arbitrary-precision, heap-managed) instead - the "safest" default, at
a real perf cost the user explicitly accepted. Scoped narrowly: a
literal still inside a TaggedUnion/TypeVar context (`x: str|i32 = 5`,
generic unification) keeps defaulting to i32, since forcing RCClass int
there would break ordinary scalar-union code.

This one-line default change had a large blast radius (275 of 1921
tests initially failed) because it exposed several previously-masked
compiler gaps - fixed each:

- Pointer subscript ASSIGNMENT (buf[0] = x) lowered its index/value
  operands with expected_type=None, unlike the read side, which already
  special-cased the index as usize. Accounted for ~90% of the blast
  radius (int.py's own buf[0] = ... broke immediately).
- Slice bounds (s[0:5]) deliberately lower with no expected type (to
  dodge an @inline-splice bug) - a bare literal bound now special-cases
  around that instead of losing its isize widening.
- Binary-op/AugAssign/comparison literal hinting (v + 10, v += 10,
  1 < a) fell back to no hint at all beside a non-scalar operand -
  previously masked because i32 happened to be both the literal default
  AND usually what the dunder/comparison partner expected. Now peeks at
  the sibling operand's own dunder parameter type (or, for `1 < a`,
  the comparison partner's type) first. New _peek_single_dunder_param_type
  helper, Overload-aware (filters to the one Scalar-typed candidate).
- Bare generic-call argument inference processed arguments in strict
  left-to-right order, letting a weak literal-default binding conflict
  with (or starve) a stronger sibling argument's binding. Now three
  tiers: concrete/typed args first, bare literals next, lambdas last
  (they have no signal of their own - PLAN_LAMBDA.md's eager lowering).
- _desugar_multi_assign's synthetic temp (`x = y = 99`) lost an
  existing target's own type, defaulting the literal to int even when
  reassigning an already-i32 local - now hints from a pre-existing
  target's Scalar type when present.
- _emit_generic_call typed its dest as `expected_type or
  monomorphized.return_type` - trusting the caller's expected type
  outright, which used to make operand.type == expected_type true BY
  CONSTRUCTION and skip _lower_expr's real coercion/mismatch check
  entirely. A genuine divergence (now possible: generic type-param
  inference is independent of the caller's own context) went completely
  undetected, reaching emitter_c.py with dest lying about its own type
  (confirmed: raw "incompatible pointer to integer conversion" in the
  generated C, un-caught by MetalPy's own checker). Fixed to type dest
  as the REAL monomorphized return type, with an explicit override for
  _lower_class_generic_method_call (e.g. Result.Ok), which already
  validates expected_type up front and can trust it.

Also fixed real pre-existing compiler bugs found along the way:
- emitter_c.py emitted CEnum typedefs AFTER interface/RCClass vtable
  struct bodies, so an enum type used BY VALUE (not pointer) in a
  vtable slot produced "unknown type name" - unrelated to this change
  directly, hit while wiring the earlier decode_lossy(errors: DecodeErrors)
  feature's vtable slot.

Not fixed, deliberately: _natural_literal_type (type_resolver.py) is
documented to mirror _expr_Constant's literal default but was left
stale on purpose - bringing it in sync surfaces a REAL, GENERAL,
PRE-EXISTING RC leak (confirmed reproducible on a clean master
checkout, no int/generics involved at all: a discarded call's fresh RC
argument leaks whenever its own Result gets propagated via an early
return instead of falling through normally - _emit_or_throw's
propagate branch skips the statement's own pending-temp release).
Flagged as a separate follow-up task, not fixed here.

Regression tests: codecs_test.py's own new decode_lossy coverage
(separate commit), lowering_test.py's new GenericCallDestTypeTests
(revert-verified to catch the dest-typing miscompile), plus several
pre-existing tests updated to use explicit i32(...)/int(...) where they
relied on the old i32 literal default (emitter_c_test.py's CallableTests,
lowering_test.py's GenericCallDefaultParameterTests and lambda-capture
test).

Verified: full 1921/1923-test suite green on clang, MSVC, and WSL/gcc
throughout; every individual fix confirmed via a real standalone
repro and a revert-then-confirm-fails pass before landing.

+++ emitter_c_test.py
@@ -16967,7 +16967,15 @@ def apply[T,K]( x: T, key: Ptr[Callable[[T],K]] ) -> K:
 	return key( x )
 
 def main() -> i32:
-	result: i32 = apply( 5, key = lambda v: v )
+	# i32(5), not a bare 5 - a bare int literal's own natural type is now
+	# builtins.int (arbitrary-precision), not i32 (see lowering.py's
+	# _expr_Constant), and nothing else in this call links T back to K
+	# (they're independent type params in apply's own signature - only the
+	# lambda's own trivial identity body happens to make T==K here, which
+	# the type-param inference has no way to see ahead of lowering the
+	# lambda) - so T needs to be pinned to i32 explicitly for this to
+	# still exercise the eager-lambda-lowering path it's actually testing
+	result: i32 = apply( i32( 5 ), key = lambda v: v )
 	with compiler.wrap_arithmetic:
 		diff: i32 = result - 5
 	return diff
